refactor(types): split api-keys.ts into entities/api-key.ts
